accumulator: provide more sensical API
Accumulators are not meant to be created by a central authority with the
private key. Rather, the private key should not be known by any party,
and only the public key should be used to create accumulators. For
example, see the applications in the paper:

One-Way Accumulators: A Decentralized Alternative to Digital Signatures
by Benaloh and de Mare
http://www.michaeldemare.com/pubs/owa.pdf

Zerocoin also uses accumulators in this way.